Add files via upload
Here Arduino is the heart and brain of this system. Arduino controls the complete voting processes like reading button, incrementing vote value, generating a result, and sending vote and result in LCD Display.Here we have added five buttons that are assigned for team A, team B, team C, team D, and the last button is used for calculating or displaying results.
In this project, i.e Smart Electronic Voting Machine Using Arduino, we have used four pushbuttons for four different candidates who are taking part in the election. We can increase the number of the candidate as per requirement. When any voter press any of four buttons then respecting voting value will increment by one each time. After the whole voting process, the result button can be pressed to display the result.
Smart Electronic Voting Machine will take the inputs from the user in the form of the votes and will record them and accordingly will display the name of the party that wins and also returns the difference between the votes of the two parties.
In this project, three push buttons are used in which two are for recording the votes and one for displaying the result.